fix: Use toInt() for Byte bitwise AND operation
Fixed compilation error where bitwise AND operator was used on Byte type.

## Issue
- Error: "Unresolved reference 'and'" on line 197
- Kotlin Byte type doesn't have infix `and` operator
- Must convert to Int before bitwise operations

## Root Cause
Byte values in Kotlin don't support the `and` infix operator directly.
The bitwise AND operation requires Int type.

## Fix
Changed:
```kotlin
(rangeBytes[i] and mask[i]) == (testBytes[i] and mask[i])
```

To:
```kotlin
(rangeBytes[i].toInt() and mask[i].toInt()) == (testBytes[i].toInt() and mask[i].toInt())
```

## Technical Details
- Kotlin Byte: Signed 8-bit integer (-128 to 127)
- Bitwise operations: Require Int conversion
- The `and` infix function is only available for Int, Long, etc.

## Impact
- Fixes compilation error
- IP range matching now works correctly
- No functional change in logic

+++ app/src/main/kotlin/com/simplexray/an/protocol/routing/AdvancedRouter.kt
@@ -194,7 +194,7 @@ class AdvancedRouter {
                 val mask = createMask(prefixLength, rangeBytes.size)
 
                 rangeBytes.indices.all { i ->
-                    (rangeBytes[i] and mask[i]) == (testBytes[i] and mask[i])
+                    (rangeBytes[i].toInt() and mask[i].toInt()) == (testBytes[i].toInt() and mask[i].toInt())
                 }
             } catch (e: Exception) {
                 false
